Make it abstract for consistency
