--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/IRazorConfigurationService.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Threading.Tasks;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer
-{
-    internal interface IRazorConfigurationService
-    {
-        Task<RazorLSPOptions> GetLatestOptionsAsync();
-    }
-}
